Moved the DataTypes to the club.psychose.library.ibo.core package.
